Ftr/customized ranking (#38)
* goggles parser

* support weird urls in goggle

* quite significant speedboost during indexing (approx x1.7)

* merge index explicitly in indexer

* query benchmark

* search performance improvements primarily by not using hash during search

* turn goggle into tantivy query

* goggle benchmark

* Goggles are working!!

* fixed bug where goggle would enforce that sitename must be A and B and C etc. instead of A or B or C

* document some more goggle syntax

* less bold search highlights

+++ src/ranking/initial.rs
@@ -21,7 +21,7 @@ use chrono::Utc;
 use tantivy::collector::{ScoreSegmentTweaker, ScoreTweaker};
 use tantivy::{DocId, Score, SegmentReader};
 
-use super::signal_aggregator::SignalAggregator;
+use super::goggles::SignalAggregator;
 
 pub(crate) struct InitialScoreTweaker {
     region_count: Arc<RegionCount>,
@@ -46,7 +46,7 @@ impl InitialScoreTweaker {
 pub(crate) struct InitialSegmentScoreTweaker {
     aggregator: SignalAggregator,
     region_count: Arc<RegionCount>,
-    current_timestamp: f64,
+    current_timestamp: usize,
     selected_region: Option<Region>,
 }
 
@@ -57,7 +57,7 @@ impl ScoreTweaker<f64> for InitialScoreTweaker {
         let mut aggregator = SignalAggregator::new_like(&self.aggregator);
         aggregator.register_readers(segment_reader);
 
-        let current_timestamp = Utc::now().timestamp() as f64;
+        let current_timestamp = Utc::now().timestamp() as usize;
 
         Ok(InitialSegmentScoreTweaker {
             aggregator,
